requirements.txt
openai
IPython
python-dotenv
